ai-service/requirements.txt
tensorflow
matplotlib
numpy
Pillow
streamlit
opencv-python
torch
torchvision

# run:
#   cd "D:\Final Project\When-will-banana-die\ai-service\training"
#   streamlit run app.py
